Add confirm delete

+++ src/views/JobView.vue
@@ -19,9 +19,13 @@ const state = reactive({
 
 const deleteJob = async () =>{
   try {
+    const confirm = window.confirm('Are you sure you want to delete this job?');
+    if (confirm){
     await axios.delete(`/api/jobs/${jobId}`);
     toast.success('Job Deleted Successfully');
     router.push('/jobs');
+    }
+
   } catch (error) {
     console.error('Error deleting job', error);
     toast.error('Job not deleted');
